main.js: extract reusable code to frontend.js
